+++ src/logger.ts
@@ -1,0 +1,4 @@
+import * as Logger from 'js-logger'
+Logger.useDefaults()
+Logger.setLevel(Logger.OFF)
+export { Logger }
