+++ src/proposal.ts
@@ -1,7 +1,8 @@
 import gql from 'graphql-tag'
 import { Observable, of } from 'rxjs'
-import { Arc, IApolloQueryOptions, Logger } from './arc'
+import { Arc, IApolloQueryOptions } from './arc'
 import { DAO } from './dao'
+import { Logger } from './logger'
 import { Operation, sendTransaction } from './operation'
 import { IRewardQueryOptions, IRewardState, Reward } from './reward'
 import { IStake, IStakeQueryOptions, Stake } from './stake'
